Factor out duplicate code.
