+++ src/lib.rs
@@ -17,15 +17,16 @@
 //!
 //! ```
 //! use aarch64_paging::{
-//!     idmap::IdMap,
+//!     idmap::IdTranslation,
 //!     paging::{Attributes, MemoryRegion},
+//!     Mapping,
 //! };
 //!
 //! const ASID: usize = 1;
 //! const ROOT_LEVEL: usize = 1;
 //!
 //! // Create a new page table with identity mapping.
-//! let mut idmap = IdMap::new(ASID, ROOT_LEVEL);
+//! let mut idmap = Mapping::new(IdTranslation, ASID, ROOT_LEVEL);
 //! // Map a 2 MiB region of memory as read-only.
 //! idmap.map_range(
 //!     &MemoryRegion::new(0x80200000, 0x80400000),
@@ -44,7 +45,10 @@ pub mod paging;
 
 extern crate alloc;
 
+#[cfg(target_arch = "aarch64")]
+use core::arch::asm;
 use core::fmt::{self, Display, Formatter};
+use paging::{Attributes, MemoryRegion, RootTable, Translation};
 
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub struct AddressRangeError;
@@ -52,5 +56,150 @@ pub struct AddressRangeError;
 impl Display for AddressRangeError {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         write!(f, "AddressRangeError")
+    }
+}
+
+/// Manages a level 1 page-table.
+///
+/// Mappings should be added with [`map_range`](Self::map_range) before calling
+/// [`activate`](Self::activate) to start using the new page table. To make changes which may
+/// require break-before-make semantics you must first call [`deactivate`](Self::deactivate) to
+/// switch back to a previous static page table, and then `activate` again after making the desired
+/// changes.
+///
+/// # Example
+///
+/// ```
+/// use aarch64_paging::{
+///     idmap::IdTranslation,
+///     paging::{Attributes, MemoryRegion},
+///     Mapping,
+/// };
+///
+/// const ASID: usize = 1;
+/// const ROOT_LEVEL: usize = 1;
+///
+/// // Create a new page table with identity mapping.
+/// let mut idmap = Mapping::new(IdTranslation, ASID, ROOT_LEVEL);
+/// // Map a 2 MiB region of memory as read-write.
+/// idmap.map_range(
+///     &MemoryRegion::new(0x80200000, 0x80400000),
+///     Attributes::NORMAL | Attributes::NON_GLOBAL | Attributes::EXECUTE_NEVER,
+/// ).unwrap();
+/// // Set `TTBR0_EL1` to activate the page table.
+/// # #[cfg(target_arch = "aarch64")]
+/// idmap.activate();
+///
+/// // Write something to the memory...
+///
+/// // Restore `TTBR0_EL1` to its earlier value while we modify the page table.
+/// # #[cfg(target_arch = "aarch64")]
+/// idmap.deactivate();
+/// // Now change the mapping to read-only and executable.
+/// idmap.map_range(
+///     &MemoryRegion::new(0x80200000, 0x80400000),
+///     Attributes::NORMAL | Attributes::NON_GLOBAL | Attributes::READ_ONLY,
+/// ).unwrap();
+/// # #[cfg(target_arch = "aarch64")]
+/// idmap.activate();
+/// ```
+#[derive(Debug)]
+pub struct Mapping<T: Translation + Clone> {
+    root: RootTable<T>,
+    #[allow(unused)]
+    asid: usize,
+    #[allow(unused)]
+    previous_ttbr: Option<usize>,
+}
+
+impl<T: Translation + Clone> Mapping<T> {
+    /// Creates a new page table with the given ASID, root level and translation mapping.
+    pub fn new(translation: T, asid: usize, rootlevel: usize) -> Self {
+        Self {
+            root: RootTable::new(translation, rootlevel),
+            asid,
+            previous_ttbr: None,
+        }
+    }
+
+    /// Activates the page table by setting `TTBR0_EL1` to point to it, and saves the previous value
+    /// of `TTBR0_EL1` so that it may later be restored by [`deactivate`](Self::deactivate).
+    ///
+    /// Panics if a previous value of `TTBR0_EL1` is already saved and not yet used by a call to
+    /// `deactivate`.
+    #[cfg(target_arch = "aarch64")]
+    pub fn activate(&mut self) {
+        assert!(self.previous_ttbr.is_none());
+
+        let mut previous_ttbr;
+        unsafe {
+            // Safe because we trust that self.root.to_physical() returns a valid physical address
+            // of a page table, and the `Drop` implementation will reset `TTRB0_EL1` before it
+            // becomes invalid.
+            asm!(
+                "mrs   {previous_ttbr}, ttbr0_el1",
+                "msr   ttbr0_el1, {ttbrval}",
+                "isb",
+                ttbrval = in(reg) self.root.to_physical().0 | (self.asid << 48),
+                previous_ttbr = out(reg) previous_ttbr,
+                options(preserves_flags),
+            );
+        }
+        self.previous_ttbr = Some(previous_ttbr);
+    }
+
+    /// Deactivates the page table, by setting `TTBR0_EL1` back to the value it had before
+    /// [`activate`](Self::activate) was called, and invalidating the TLB for this page table's
+    /// configured ASID.
+    ///
+    /// Panics if there is no saved `TTRB0_EL1` value because `activate` has not previously been
+    /// called.
+    #[cfg(target_arch = "aarch64")]
+    pub fn deactivate(&mut self) {
+        unsafe {
+            // Safe because this just restores the previously saved value of `TTBR0_EL1`, which must
+            // have been valid.
+            asm!(
+                "msr   ttbr0_el1, {ttbrval}",
+                "isb",
+                "tlbi  aside1, {asid}",
+                "dsb   nsh",
+                "isb",
+                asid = in(reg) self.asid << 48,
+                ttbrval = in(reg) self.previous_ttbr.unwrap(),
+                options(preserves_flags),
+            );
+        }
+        self.previous_ttbr = None;
+    }
+
+    /// Maps the given range of virtual addresses to the identical physical addresses with the given
+    /// flags.
+    ///
+    /// This should generally only be called while the page table is not active. In particular, any
+    /// change that may require break-before-make per the architecture must be made while the page
+    /// table is inactive. Mapping a previously unmapped memory range may be done while the page
+    /// table is active.
+    pub fn map_range(
+        &mut self,
+        range: &MemoryRegion,
+        flags: Attributes,
+    ) -> Result<(), AddressRangeError> {
+        self.root.map_range(range, flags)?;
+        #[cfg(target_arch = "aarch64")]
+        unsafe {
+            // Safe because this is just a memory barrier.
+            asm!("dsb ishst");
+        }
+        Ok(())
+    }
+}
+
+impl<T: Translation + Clone> Drop for Mapping<T> {
+    fn drop(&mut self) {
+        if self.previous_ttbr.is_some() {
+            #[cfg(target_arch = "aarch64")]
+            self.deactivate();
+        }
     }
 }
